Refactor after PR review and made GooglePurchaseInfo internal
