+++ examples/purchase-tester/build.gradle
@@ -72,4 +72,5 @@ dependencies {
 
     implementation "androidx.navigation:navigation-fragment-ktx:2.5.2"
     implementation "androidx.navigation:navigation-ui-ktx:2.5.2"
+    implementation project(path: ':feature:google')
 }
