+++ feature/google/src/main/java/com/revenuecat/purchases/google/GoogleStoreProduct.kt
@@ -1,8 +1,12 @@
-package com.revenuecat.purchases.models
+package com.revenuecat.purchases.google
 
 import android.os.Parcelable
 import com.android.billingclient.api.ProductDetails
 import com.revenuecat.purchases.ProductType
+import com.revenuecat.purchases.models.Price
+import com.revenuecat.purchases.models.PurchaseInfo
+import com.revenuecat.purchases.models.PurchaseOption
+import com.revenuecat.purchases.models.StoreProduct
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import kotlinx.parcelize.RawValue
@@ -24,7 +28,7 @@ data class GoogleStoreProduct(
         get() = if (type == ProductType.SUBS && defaultOption != null) {
             defaultOption.purchaseInfo
         } else {
-            GooglePurchaseInfo.NotSubscription(
+            com.revenuecat.purchases.google.GooglePurchaseInfo.InAppProduct(
                 productId,
                 productDetails
             )
