+++ feature/google/src/main/java/com/revenuecat/purchases/google/purchaseOptionConversions.kt
@@ -1,9 +1,6 @@
 package com.revenuecat.purchases.google
 
 import com.android.billingclient.api.ProductDetails
-import com.revenuecat.purchases.ProductType
-import com.revenuecat.purchases.models.GooglePurchaseInfo
-import com.revenuecat.purchases.models.GooglePurchaseOption
 
 fun ProductDetails.SubscriptionOfferDetails.toPurchaseOption(productId: String, productDetails: ProductDetails): GooglePurchaseOption {
     val pricingPhases = pricingPhases.pricingPhaseList.map { it.toRevenueCatPricingPhase() }
