+++ public/src/main/java/com/revenuecat/purchases/models/StoreProduct.kt
@@ -66,3 +66,5 @@ interface StoreProduct : Parcelable {
     )
     val sku: String
 }
+
+internal fun StoreProduct.getPurchaseInfo2() = Unit
